Added CurrencyModel definition for persistence. Added Enum that difin…

+++ src/main/resources/db/migration/V001__setup.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS currency_transaction(
     origin_currency         VARCHAR(3)           NOT NULL,
     destination_currency    VARCHAR(3)           NOT NULL,
     origin_amount           DOUBLE               NOT NULL,
-    conversion_tax          DOUBLE               NOT NULL,
+    conversion_tax          LONG                 NOT NULL,
     created_at              TIMESTAMP,
     updated_at              TIMESTAMP,
     deleted_at              TIMESTAMP            DEFAULT NULL,
